Finished: learn how to use async function to produce relative synchronize behavior, see sigVerifyTest.js for example.
