+++ public/javascripts/objects.js
@@ -4,90 +4,6 @@ to a specific deployment url, using a specific address.*/
 /*User needs to specify which contract to deploy, and pass in the 
 parameters used in the constructor of contracts.*/
 
-function contractDeployer(_url_toDeploy,_senderAddr){
-    var Web3=require('web3');
-    var web3=new Web3(new Web3.providers.HttpProvider(_url_toDeploy));
-    web3.eth.defaultAccount=_senderAddr;
-
-    this.url_toDeploy=_url_toDeploy;
-    console.log("_senderAddr",_senderAddr);
-    this.senderAddr=_senderAddr;
-
-    if(this.url_toDeploy==undefined)
-        url_toDeploy="http://localhost:8545";
-
-   
-   this.RC_deploy=function(){
-       console.log("Deploying an RC");
-       return this.deploy_template("RC");
-}
-    this.PPR_deploy=function(patientID,providerID){
-        console.log("Deploying an PPR");
-    return this.deploy_template("PPR",patientID.toString(),providerID.toString());
-}
-    this.SC_deploy=function(RegistrarID,RegistrarEthAddr){
-        console.log("Deploying an SC");
-    return this.deploy_template("SC",RegistrarID.toString(),RegistrarEthAddr);
-}
-
-    this.generateTemplate=function(sc_type){
-      
-        var format_src;
-        switch(sc_type){
-            case "RC": format_src='../../build/contracts/RC.json'; break;
-            case "PPR": format_src='../../build/contracts/PPR.json'; break;
-            case "SC": format_src='../../build/contracts/SC.json'; break;
-            default: console.log('Unknown contract type.'); return undefined;
-        }
-
-        var contract_format=require(format_src);
-        // var gasEstimate= web3.eth.estimateGas({data:contract_format.bytecode});
-        
-        
-        var contract_generator=new web3.eth.Contract(contract_format.abi);
-
-        return {template:contract_generator,format:contract_format};
-    }
-
-    this.deploy_template=function(sc_type,param1,param2){
-        
-        var contract_generator=this.generateTemplate(sc_type);
-     
-        
-        console.log("Deploying contract.");
-        var gasEstimate=1000000;
-        //Deploy a new contract on rpc.
-        console.log(this.senderAddr);
-
-        var contract_instance=contract_generator.template.deploy(
-        {
-            data:contract_generator.format.bytecode,
-            arguments:[param1,param2]
-        }).send({gas:gasEstimate,gasPrice:'0',from:this.senderAddr},function(err,transactionHash){
-            if(!err)
-                {
-                    console.log("Contract Deployed.");
-                    
-                    console.log("TxHash:",transactionHash);
-                }
-            else    
-                console.log(err);
-        });
-        //After the call above, the contract address is still undefined, as it may not have been mined yet.
-        ///
-
-
-        //We refresh the contract_instance to be the actual contract on chain.
-        // var receipt=web3.eth.getTransactionReceipt(contract_instance.transactionHash).contractAddress;
-        // console.log(receipt);
-        // contract_instance=contract_generator.template.at(address);
-        // console.log(contract_instance);
-        ///
-
-        return contract_instance;
-    }
-}
-
 function DbGatekeeper(){
     
     this.PPR_List=[];
